# Include fast convolution based dice probabilities