A faster way to compute the dice rolls is using the [convolution of the probability mass functions](https://en.wikipedia.org/wiki/Convolution_of_probability_distributions).

Which an intimidating mathy way of saying something that isn't too complicated.

The intuition is to count the number of ways to create each sum. For example, in a 1d3 roll there is 1 way to create each of 1, 2, or 3:
```
1d3 = {1:1, 2:1, 3:1}
```

A 2d3 roll, combines two 1d3 rolls:
```
(roll1, roll2) = roll1 + roll2
(1, 1) = 2
(1, 2) = 3
(1, 3) = 4
(2, 1) = 3
(2, 2) = 4
(2, 3) = 5
(3, 1) = 4
(3, 2) = 5
(3, 3) = 6

# Then count the number of ways to make each sum:
2d3 = {2: 1, 3:2, 4:3, 5:2, 6:1}
```

A 3d3 roll could combine a 1d3 and 2d3 roll:
```
1d3 = {1:1, 2:1, 3:1}
2d3 = {2: 1, 3:2, 4:3, 5:2, 6:1}

# We need to take the counts into consideration

(roll1: c1, roll2: c2) = sum: c1 * c2
(2:1, 1:1) = 3: 1*1
(2:1, 2:1) = 4: 1*1
(2:1, 3:1) = 5: 1*1
(3:2, 1:1) = 4: 2*1
(3:2, 2:1) = 5: 2*1
(3:2, 3:1) = 6: 2*1
(4:3, 1:1) = 5: 3*1
(4:3, 2:1) = 6: 3*1
(4:3, 3:1) = 7: 3*1
(5:2, 1:1) = 6: 2*1
(5:2, 2:1) = 7: 2*1
(5:2, 3:1) = 8: 2*1
(6:1, 1:1) = 7: 1*1
(6:1, 2:1) = 8: 1*1
(6:1, 3:1) = 9: 1*1

3d3 = {
  3: 1,  # 1
  4: 3,  # 1+2
  5: 6,  # 1+2+3
  6: 7,  # 2+3+2
  7: 6,  # 3+2+1
  8: 3,  # 2+1
  9: 1   # 1
}
```