@Lanzaa 2021-09-09
The timings on my machine are as follows:

roll|old (ms)|conv(ms)
-|-|-
14d6|30|1
20d6|260|1
30d6|1703|2
14d10|2330|1
20d10|37031|4
30d10|?|16
50d10|?|17
100d5|?|48
200d5|?|133